remove jenkins & travis, break out ViewOperation along conventions
