--- .travis.yml
@@ -1,26 +1,0 @@
-language: generic
-
-
-matrix:
-  allow_failures:
-    - os: linux
-  include:
-    - os: linux
-      dist: xenial
-      sudo: required
-      script:
-        - docker-compose -f couchdb.yml -f swift.yml up --abort-on-container-exit
-      after_script: 
-        - docker-compose -f couchdb.yml -f swift.yml down -v --rmi local
-    - os: osx
-      language: objective-c
-      osx_image: xcode10.2
-      before_script:
-        - brew install couchdb
-        - brew services start couchdb
-      script:
-        - swift build && swift test
-
-services:
-    - couchdb
-    - docker
