+++ Source/SwiftCloudant/Operations/Views/ViewOperation/Stale.swift
@@ -1,0 +1,36 @@
+// Stale.swift
+//  
+//  SwiftCloudant
+//
+//  Created by Dan Burkhardt on 8/24/23.
+//  Copyright © 2023 IBM Corporation. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+//  except in compliance with the License. You may obtain a copy of the License at
+//    http://www.apache.org/licenses/LICENSE-2.0
+//  Unless required by applicable law or agreed to in writing, software distributed under the
+//  License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+//  either express or implied. See the License for the specific language governing permissions
+//  and limitations under the License.
+//
+
+
+import Foundation
+
+// This Can't be defined in the extension or the protocol defining outside for now.
+/**
+ Acceptable values for allowing stale views with the `stale` property. To disallow stale views use the default `stale=nil`.
+ */
+public enum Stale: CustomStringConvertible {
+    /** Allow stale views. */
+    case ok
+    /** Allow stale views, but update them immediately after the request. */
+    case updateAfter
+    
+    public var description: String {
+        switch self {
+        case .ok: return "ok"
+        case .updateAfter: return "update_after"
+        }
+    }
+}
