+++ Source/SwiftCloudant/Operations/Views/ViewOperation/ViewOperation.swift
@@ -17,25 +17,6 @@
 
 import Foundation
 
-
-// This Can't be defined in the extension or the protocol defining outside for now.
-/**
- Acceptable values for allowing stale views with the `stale` property. To disallow stale views use the default `stale=nil`.
- */
-public enum Stale: CustomStringConvertible {
-    /** Allow stale views. */
-    case ok
-    /** Allow stale views, but update them immediately after the request. */
-    case updateAfter
-    
-    public var description: String {
-        switch self {
-        case .ok: return "ok"
-        case .updateAfter: return "update_after"
-        }
-    }
-}
-
 /**
  Denotes an operation that performs actions on a view.
  Some operations are not strictly a view externally, but internally
@@ -134,100 +115,4 @@ public protocol ViewOperation : CouchDatabaseOperation {
      */
     var rowHandler: (([String: Any]) -> Void)? { get }
     
-}
-
-public extension ViewOperation {
-    
-    func processResponse(json: Any) {
-        if let json = json as? [String: Any] {
-            let rows = json["rows"] as! [[String: Any]]
-            for row: [String: Any] in rows {
-                self.rowHandler?(row)
-            }
-        }
-    }
-    
-    var method: String {
-        get {
-            if keys != nil {
-                return "POST"
-            } else {
-                return "GET"
-            }
-        }
-    }
-    
-    /** 
-    Generates parameters for the following properties
- 
-    * descending
-    * startKeyDocId
-    * endKeyDocId
-    * inclusiveEnd
-    * limit
-    * skip
-    * includeDocs
-    * conflicts
-     
-     
-    - Note: Implementing types *have* to add parameters which use the `associatedtype` `ViewParameter`
-    */
-    func makeParams() -> [String : String]{
-        var items: [String: String] = [:]
-        
-        if let descending = descending {
-            items["descending"] = "\(descending)"
-        }
-        
-        if let startKeyDocId = startKeyDocumentID {
-            items["startkey_docid"] = startKeyDocId
-        }
-        
-        if let endKeyDocId = endKeyDocumentID {
-            items["endkey_docid"] = "\(endKeyDocId)"
-        }
-        
-        if let inclusiveEnd = inclusiveEnd {
-            items["inclusive_end"] = "\(inclusiveEnd)"
-        }
-        
-        if let limit = limit {
-            items["limit"] = "\(limit)"
-        }
-        
-        if let skip = skip {
-            items["skip"] = "\(skip)"
-        }
-        
-        if let includeDocs = includeDocs {
-            items["include_docs"] = "\(includeDocs)"
-        }
-        
-        if let stale = stale {
-            items["stale"] = "\(stale)"
-        }
-        
-        if let conflicts = conflicts {
-            items["conflicts"] = "\(conflicts)"
-        }
-        
-        if let includeLastUpdateSequenceNumber = includeLastUpdateSequenceNumber {
-            items["update_seq"] = "\(includeLastUpdateSequenceNumber)"
-        }
-        
-        return items
-    }
-    
-    func jsonValue(for key: Any) throws -> String {
-        if JSONSerialization.isValidJSONObject(key) {
-            let keyJson = try JSONSerialization.data(withJSONObject: key)
-            return String(data: keyJson, encoding: .utf8)!
-        } else if key is String {
-            // we need to quote JSON primitive strings
-            return "\"\(key)\""
-        } else {
-            // anything else we just try as stringified JSON value
-            return "\(key)"
-        }
-    }
 }
